HDDS-11283. Refactor KeyValueStreamDataChannel.class to avoid spurious IDE build issues
