+++ hadoop-hdds/container-service/src/main/java/org/apache/hadoop/ozone/container/keyvalue/impl/KeyValueStreamDataChannel.java
@@ -54,91 +54,6 @@ public class KeyValueStreamDataChannel extends StreamDataChannelBase {
   public static final Logger LOG =
       LoggerFactory.getLogger(KeyValueStreamDataChannel.class);
 
-  /**
-   * Keep the last {@link Buffers#max} bytes in the buffer
-   * in order to create putBlockRequest
-   * at {@link #closeBuffers(Buffers, WriteMethod)}}.
-   */
-  static class Buffers {
-    private final Deque<ReferenceCountedObject<ByteBuffer>> deque
-        = new LinkedList<>();
-    private final int max;
-    private int length;
-
-    Buffers(int max) {
-      this.max = max;
-    }
-
-    private boolean isExtra(int n) {
-      return length - n >= max;
-    }
-
-    private boolean hasExtraBuffer() {
-      return Optional.ofNullable(deque.peek())
-          .map(ReferenceCountedObject::get)
-          .filter(b -> isExtra(b.remaining()))
-          .isPresent();
-    }
-
-    /**
-     * @return extra buffers which are safe to be written.
-     */
-    Iterable<ReferenceCountedObject<ByteBuffer>> offer(
-        ReferenceCountedObject<ByteBuffer> ref) {
-      final ByteBuffer buffer = ref.retain();
-      LOG.debug("offer {}", buffer);
-      final boolean offered = deque.offer(ref);
-      Preconditions.checkState(offered, "Failed to offer");
-      length += buffer.remaining();
-
-      return () -> new Iterator<ReferenceCountedObject<ByteBuffer>>() {
-        @Override
-        public boolean hasNext() {
-          return hasExtraBuffer();
-        }
-
-        @Override
-        public ReferenceCountedObject<ByteBuffer> next() {
-          final ReferenceCountedObject<ByteBuffer> polled = poll();
-          length -= polled.get().remaining();
-          Preconditions.checkState(length >= max);
-          return polled;
-        }
-      };
-    }
-
-    ReferenceCountedObject<ByteBuffer> poll() {
-      final ReferenceCountedObject<ByteBuffer> polled
-          = Objects.requireNonNull(deque.poll());
-      RatisHelper.debug(polled.get(), "polled", LOG);
-      return polled;
-    }
-
-    ReferenceCountedObject<ByteBuf> pollAll() {
-      Preconditions.checkState(!deque.isEmpty(), "The deque is empty");
-      final ByteBuffer[] array = new ByteBuffer[deque.size()];
-      final List<ReferenceCountedObject<ByteBuffer>> refs
-          = new ArrayList<>(deque.size());
-      for (int i = 0; i < array.length; i++) {
-        final ReferenceCountedObject<ByteBuffer> ref = poll();
-        refs.add(ref);
-        array[i] = ref.get();
-      }
-      final ByteBuf buf = Unpooled.wrappedBuffer(array).asReadOnly();
-      return ReferenceCountedObject.wrap(buf, () -> {
-      }, () -> {
-        buf.release();
-        refs.forEach(ReferenceCountedObject::release);
-      });
-    }
-
-    void cleanUpAll() {
-      while (!deque.isEmpty()) {
-        poll().release();
-      }
-    }
-  }
-
   interface WriteMethod {
     int applyAsInt(ByteBuffer src) throws IOException;
   }
@@ -184,7 +99,7 @@ public class KeyValueStreamDataChannel extends StreamDataChannelBase {
 
   private static void writeFully(ByteBuffer b, WriteMethod writeMethod)
       throws IOException {
-    for (; b.remaining() > 0;) {
+    while (b.remaining() > 0) {
       final int written = writeMethod.applyAsInt(b);
       if (written <= 0) {
         throw new IOException("Unable to write");
